Remove `utxoSuitableForCollateral` from `SelectionConstraints`.
The coin selection library is currently responsible for filtering the
available UTxO set for entries that are suitable for use as collateral.

However, this responsibility is actually still handled by the wallet, as
the wallet must provide a definition of the following function:

```
    , utxoSuitableForCollateral
        :: (TxIn, TxOut) -> Maybe Coin
        -- ^ Indicates whether an individual UTxO entry is suitable for use as
        -- a collateral input. This function should return a 'Coin' value if
        -- (and only if) the given UTxO is suitable for use as collateral.
```

The coin selection library merely *applies* this function as a filter on
the set of available UTxO entries.

The requirement to provide the `utxoSuitableForCollateral` function
means that coin selection must depend on the `TxOut` type (as collateral
suitability is a function of the `Address` type)

Our goal is to evolve the coin selection library so that it does not
depend on wallet-specific types like `TxOut`, `Address` and `UTxO`.

This commit removes the `utxoSuitableForCollateral` function from
`SelectionConstraints`, and makes the wallet responsible for filtering
the `utxoAvailableForCollateral` set.