Drop check of wdrl and cert witnesses
I believe the current logic will overestimate the required number of witnesses if either:

1. A withdrawal key is identical to a input payment key
2. A certificate key is identical to a input payment key
3. There are multiple certs with the same key
4. There's a cert and withdrawal with the same key
5. Multiple withdrawals with the same key (absurd, not sure if technically possible)

In particular (4) could be a common occurrence: users re-delegating and the wallet automatically
withdrawing their rewards at the same time. An integration test was
introduced with the previous commit.

This commit is the simplest fix for the test introduced in the previous
commit.