Use `(TxIn, Address`) as a unique input identifier within coin selection.
ADP-1448

We eventually want to change the internal coin selection library so that it:
- does not depend on the following wallet-specific types:
    - `TxIn`
    - `TxOut`
    - `Address`
    - `UTxO`
- produces selections of the form:
    ```haskell
    data Selection input = Selection
        { inputs     = [(input, TokenBundle)]
        , collateral = [(input, Coin       )]
        , ...
        }
    ```
    (here `input` is a **_type parameter_** that **_uniquely_** identifies a
    selected input)

However, the **_wallet_** currently expects to manipulate selections of the
following form:

```haskell
data Selection = Selection
    { inputs =     [(TxIn, TxOut)]
    , collateral = [(TxIn, TxOut)]
    , ...
    }
```

Since (for the time being) we wish to avoid making major changes to the wallet
codebase, we need a way to translate a `Selection` produced by the internal
algorithm to a wallet-friendly `Selection`.

One way to achieve this is to instantiate type `input` as `(TxIn, Address)`, so
that our internal selection type is instantiated to:

```haskell
data Selection = Selection
    { inputs     = [((TxIn, Address), TokenBundle)]
    , collateral = [((TxIn, Address), Coin       )]
    , ...
    }
```

With this substitution, it's trivial to convert from internal `Selection`
values to wallet-friendly `Selection` values, because the choice of concrete
type for `input` has all the context necessary to reconstruct a `(TxIn, TxOut)`
pair.

This commit **_doesn't_** introduce any new type parameters, but instead makes
an evolutionary step **_towards_** the goal outlined above.

This commit:

  - Introduces a (temporary) type synonym `InputId = (TxIn, Address)`.

  - Amends core coin selection modules to use `InputId` as the unique
    identifier for inputs.

  - Amends core coin selection modules so that they no longer depend on
    the `TxOut` and `UTxO` types.

  - Amends the `readWalletUTxOIndex` function to perform the translation
    from `UTxO` (equivalent to `Map TxIn TxOut`) to `Map InputId TokenBundle`.

  - Amends the wallet-facing `Cardano.Wallet.CoinSelection` module to
    perform the reverse translation from `(InputId, TokenBundle)` pairs
    to `(TxIn, TxOut)` pairs.

A **future commit** will replace the `InputId` type synonym with an `input`
type parameter on the following types:

  - `Selection`
  - `SelectionParams`
  - `UTxOIndex`
  - `UTxOSelection`

It's important that the translations stated above do not add any performance
overhead.

  - The translation from `(TxIn, TxOut)` to `(InputId, TokenBundle)` within
    function `readWalletUTxOIndex` does need to traverse the UTxO, but it only
    does so a single time. When estimating a fee, this function is only
    evaluated a single time.

  - The translation from `(InputId, TokenBundle)` to `(TxIn, TxOut)` is only
    applied to the inputs that are eventually selected.