Change kind of `outputs` type parameter in `Selection{Params,Result}Of`.
Before this change, `outputs` was always instantiated in one of two ways:

    - outputs ~ [        (Address, TokenBundle)]
    - outputs ~ NonEmpty (Address, TokenBundle)

In both cases, the type of an individual output is `(Address, TokenBundle)`.
The only difference between these instantiations is the container type.

As a result of this change, the above instantiations are simplified to:

    - outputs ~ []
    - outputs ~ NonEmpty

This allows us to simplify the type signatures of many functions that
use the `SelectionParamsOf` and `SelectionResultOf` types, because we
now only need to specify the container type, and not the type of output.