Retain the original selection strategy in `onUnableToConstructChange`.
This change alters the `onUnableToConstructChange` handler within the
`prop_performSelection` property test. This handler is designed to
re-run a selection with relaxed constraints, and expects the selection
to succeed.

We change the handler so that it relaxes **only** the constraints that
really need to be relaxed, and retains all others, including the
selection strategy.