Move field `selectionStrategy` from `SelectionConstraints` to `SelectionParams`.
This field was actually in the wrong place.

Here's why:

  - Selection constraints are dependent on the protocol parameters, and
    NOT specific to a given selection.

  - Selection parameters are independent of the protocol parameters, and
    ARE specific to a given selection.

The choice of selection strategy IS specific to a given selection, and
NOT dependent on the protocol parameters, so it falls into the latter
category.

Incidentally, this change is also necessary to support the creation of
extra unit tests for the `Balance` module, since the `SelectionParams`
type is used to specify the nature of each unit test.