Correct Passphrase.Legacy implementation to make tests pass
1. Swap order of cborify and hashMaybe

This makes the legacy golden tests pass, and matches the old
implementation.

2. Fix getSalt implementation

The salt should be converted from base64. This fixes the test
"Verify passphrase" calling `checkPassphraseTestingOnly`.

The getSalt golden value was updated. With the "Verify passphrase" test
now succeeding, it seems like the previous expected salt value — "abc" —
was just a dummy value, and not the right expectation.