Remove unnecessary UTxO map conversions when constructing log values.
When recording details about the UTxO set in the log, we generally only
care about the size of the UTxO set, rather than its composition.

Therefore, we don't need to require ourselves to produce a value of type
`UTxO` in order to create a log entry: we just need to obtain the size
of the available UTxO set.

Therefore, we can avoid converting from the coin selection's internal
map of UTxO entries to values of the wallet's `UTxO` type, which
requires traversing the entire UTxO set and converting every single
entry.

We can't avoid traversing the wallet's UTxO set at least once, in order
to make a value of type `UTxOIndex`, but we can avoid doing it multiple
times.