Use explicit powers-of-two in `sizeOfCoin` boundary tests.
In the implementation of `sizeOfCoin`, we use hard-coded constants for
powers of two, which avoids computing these values on each call.

However, in the test code, we can be even clearer about our intent, by
explicitly stating which powers of two we're interested in.

For someone new to the codebase, being able to see the relationship with
the powers of two is arguably helpful for being able to comprehend why
we're testing these particular values.

At a glance, it's arguably quite hard to see the the difference between:
    - 4294967295
    - 4294967296

But it's very easy to see the difference between:
    - 2^32 - 1
    - 2^32