Rename `CreatedTransactionWithTooLongAssetName``.
Rename `CreatedTransactionWithTooLongAssetName` to `AssetNameTooLong`.

This error message implies that we've actually created a transaction.

But if an asset name is too long, then it isn't possble for us to have
created a transaction. Therefore, we can't say that we've created one.

Additionally, if an asset name is too long, then it's too long in all
contexts.

Therefore, it's enough to say that "the asset name is too long": we
don't actually need to state the context.

We already use this style in other areas. For example:

- AssetNotPresent
- WalletAlreadyExists