Extract out function `prop_extraFee_coversIncreaseToFeeRequirement`.
While smaller inline properties are perfectly fine, if the definition of
the property becomes too large, then it can be come a challenge to fit
it within the horizontal line limit, resulting in code that is squished
to the right, or code that has excessive wrapping.

This commit moves the definition of an inline property to a top-level
function, giving us more horizontal space to work with. This means that
some statements no longer need to be wrapped. We can unwrap them, which
saves us vertical space.