Use `where` clause in `prop_extraFee_coversIncreaseToFeeRequirement`.
Terms defined in `where` clauses can often be indented further to the
left than terms defined in `let` clauses. In this case, it means we no
longer need to wrap the definitions of these terms, which means we can
save vertical space, making the function more concise.

We also remove an unneeded `do`, and an unneeded `$`.