Make `burnSurplusAsFees` use `Empty` change container type.
Before this commit, this function was only ever called with a change
value of `Nothing`, and only ever returned a change value of `Nothing`.

Therefore, we can safely replace the use of `Maybe` with `Empty`, which
only has a single `Empty` constructor. This clarifies that this function
cannot manipulate change values in any way.