Remove usages of `Empty` and `Solo`.
These types were introduced because we parameterized the **container type**
for change in `TxFeeAndChange`:

```hs
data TxFeeAndChange f = TxFeeAndChange
    { fee :: Coin
    , change :: f Coin
    }
```

This required us to instantiate `f` with types of kind `* -> *`.

But later on we parameterized the type of `change` so that it can be any type
of kind `*`:

```hs
data TxFeeAndChange change = TxFeeAndChange
    { fee :: Coin
    , change :: change
    }
```

Hence we can make the following simplifications:

  - Empty Void -> ()
  - Solo  Coin -> Coin

And therefore we can remove the `Empty` and `Solo` types.

In response to review discussion:

https://github.com/input-output-hk/cardano-wallet/pull/3238#discussion_r853105496