Add field `collateralOutputs` to type `ApiDecodedTransaction`.
In the API specification we use an array type for `collateral_outputs`,
for consistency with the encoding of ordinary transaction outputs, and
enabling forward compatibility if in future more than one collateral
output is allowed in a transaction.

But we clearly mark the fact that there can only be zero or one
collateral output. Attempting to parse an `ApiDecodedTransaction` with
more than one collateral output will fail with an error.

For the moment, we populate this field with `Nothing`.