Add field `collateralOutput` to type `Tx`.
Important:

This commit breaks the DB tests, because it provides no way to write or
read collateral outputs to or from the DB. A further commit will provide
proper DB persistence for collateral outputs.

Instead, in this commit, we simply add the field, and fix up all places
that the compiler requires us to fix.

In places where we are providing mock data for tests of transactions
belonging to previous eras, we can simply assign `Nothing` to the field.

In places where we will eventually need to assign something other than
`Nothing`, we add a `TODO`, with a link to an appropriate ticket number.