Add `eraHistory` getter to `NetworkLayer`
In order to switch from `Alonzo.evaluateTransactionExecutionUnits` to
`Cardano.evaluateTransactionExecutionUnits` we need an `EraHistory`.
Because `TimeInterpreter` can internally contain an EraSummary of any
shape (forall xs), and `EraHistory` only works with a closed set of
modes, we cannot convert between the two, and thus need a new accessor
to the `NetworkLayer`.

There should in principle exist several ways to reduce this duplication
in the future.

Includes squashed review suggestions.

Co-authored-by: Jonathan Knowles <jonathan.knowles@iohk.io>