Remove metadata schema parameter from `Api.Link.getTransaction`.
This function is always called with a value of `False` for the "simple
metadata" parameter. Therefore, we can remove this parameter from the
function and use a constant, which removes the burden of including the
parameter at all call sites.

To avoid Boolean blindness, we specify the `TxMetadataDetailedSchema`
constructor explicitly (which is equivalent to simple metadata = `False`).