Replace `unit_applyTxToUTxO_loses_collateral` with stronger tests.
Before this commit, if we were to mutate the `txScriptInvalid` function
so that it erroneously returned 'True' or erroneously returned 'False',
then the model test suite would **not** fail.

To make up for this deficiency in coverage, in this commit we replace
the `unit_applyTxToUTxO_loses_collateral` test with a trio of stronger
tests:

- unit_applyTxToUTxO_scriptValidity_Valid
- unit_applyTxToUTxO_scriptValidity_Invalid
- unit_applyTxToUTxO_scriptValidity_Unknown

These tests avoid relying on the correctness of `txScriptInvalid`, and
instead specify the script validity directly.