Make it possible to show values of `IsOursIf2`.
When testing arbitrary transaction sequences, we sometimes need to apply
transactions that may or may not belong to a wallet. We use `IsOursIf2`
to simulate wallet states where only a subset of the address space
belongs to the wallet. The ability to show values of `IsOurIf2` is
extremely useful for understanding counterexamples.