Remove unused ErrAssignRedeemersPastHorizon for now
It seems we can no longer construct this error as the ledger no longer
returns the actual `PastHorizon`. This commit removes it, with the
intention that we can re-add it being explicitly rendered in our API
instead of being lumped under TranslationErrors. It would probably also
be nice to render all the other TranslationErrors explicitly.