Remove NO_POOLS env var in cluster
It was useful for making the cluster startup shorter by not waiting
multiple epochs for the pools to get stake. With babbage we can register
pools in genesis to be active immediately, so the primary use for NO_POOLS
is removed.

In alonzo or prior eras, NO_POOLS could be useful, but for simplicity
let's remove it anyway.