Add commentary to `prop_computeMinimumCoinForUTxO_shelleyBasedEra_stability`.
We further explain the purpose of this property.

If the Cardano API function `calculateMinimumUTxO` is ever changed so
that it computes a fixed point before returning its result, this
property will fail.

However, this is valuable information, as it tells us that we might be
able to revise our implementation of `computeMinimumCoinForUTxO` to
reduce the level of overestimation.

In response to review feedback.