Improve realism of generated test protocol parameters in `MinimumUTxO.Gen`.
This commit introduces generator function `genLedgerCoinOfSimilarMagnitude`,
which, when given a coin value, will generate another coin value that is
of a similar magnitude.

When generating test values of `MinimumUTxO`, we apply the generator
function `genLedgerCoinOfSimilarMagnitude` to protocol parameter values
obtained from real mainnet genesis files. This enables us to generate
more realistic values of `MinimumUTxO`.

In response to review feedback:
https://github.com/input-output-hk/cardano-wallet/pull/3368#discussion_r918912368