Update `balanceTx` delegation golden test values.
It's unclear why the error message suggests we need 2508 more lovelace
to generate change. By reducing the minUTxOValue overestimation we
should require /less/ ada for change.

It might be that the 'dummyAddress' used to be involved, and
artificially lowered the minUTxOValue reported in the error message.

It would be nice to eventually test that the lovelace quantities
reported by error messages actually are correct and useful in some
sense: e.g.  increasing the ada quantity by of an input by `shortfall`
should presumably make the operation succeed, and be the lowest amount
which does so.