Raise `UnableToConstructChange.shortfall` values for `balanceTx` golden tests.
In the case of delegation, the wallet typically calls coin selection
without any user-specified outputs.

In such cases, coin selection will use a temporary dummy output to act as
a target for change generation, defined within the `performSelectionEmpty`
function.

Although we have adjusted this function to use a valid address of
minimal length, this address is still longer than the null-length dummy
address that we were able to use when the `computeMinimumCoinForUTxO`
function did not require an address.

This increase in address length causes an increase in the estimated cost
when generating change. Therefore, any `UnableToConstructChange` errors
will report ada shortfalls that are larger in magnitude than before.

A future commit will adjust `performSelectionEmpty` so that it can use a
null-length address again, so we can minimize the cost overestimation.