Add `BoundedAddressLength` type class and instances.
This class provides a way to specify that addresses are bounded in
length for a particular key type.

In principle, for a given key type, the set of valid address sizes is
bounded in both directions: there will exist both a minimum and maximum
length.

In practice, for now, we only define the `maxLengthAddressFor` function,
which returns an address of maximum length for the given key.

Later on, if needed, we could add a `minLengthAddressFor` function to
specify minimum length addresses.

Co-Authored-By: Johannes Lund <johannes.lund@iohk.io>