Raise cost of joining and quitting in `STAKE_POOLS_{JOIN,QUIT}_01x`.
In the case of delegation, the wallet typically calls coin selection
without any user-specified outputs.

In such cases, coin selection will use a temporary dummy output to act as
a target for change generation, defined within the `performSelectionEmpty`
function.

Although we have adjusted this function to use a valid address of
minimal length, this address is still longer than the null-length dummy
address that we were able to use when the `computeMinimumCoinForUTxO`
function did not require an address.

This increase in address length causes an increase in the estimated cost
when generating change. Therefore, this also raises the cost of joining
or quitting a stake pool.

A future commit will adjust `performSelectionEmpty` so that it can use a
null-length address again, so we can minimize the cost overestimation.