Use a null address and minimal coin value within `performSelectionEmpty`.
Now that `performSelectionNonEmpty` no longer needs to validate the
minimum ada quantities of user-specified outputs (a responsibility that
has been moved to the top-level coin selection module), we are free to
choose any `Address` (even if invalid) and any non-zero `Coin` value for
the dummy output used by `performSelectionEmpty`.

In order to minimize any overestimation in cost resulting from the use
of a dummy output, we choose the shortest possible `Address` (which is a
null-length address) and the smallest possible non-zero `Coin` value.