Lower `UnableToConstructChange.shortfall` values for `balanceTx` golden tests.
The previous commit adjusted the dummy output used by function
`performSelectionEmpty` so that it uses a null-length `Address` value.

This null-length `Address` value is identical to the one we used before
we adjusted `computeMinimumCoinForUTxO` to require a valid `Address`.

However, we were also able to adjust `performSelectionNonEmpty` so that
it no longer needs to validate minimum ada quantities of user-specified
outputs, and consequently we were able to reduce the dummy output's
`Coin` value to just `Coin 1`, which has the smallest space cost of any
non-zero `Coin` value.

This results in a further reduction in cost overestimation while
constructing change, which results in smaller `shortfall` values within
`UnableToConstructChange` errors.