Revise definition of the API field `minimum_utxo_value`.
We adjust the internal definition so that it uses an address of the
minimum length.

As a result, we can now say that the value returned is an absolute
minimum w.r.t. to the wallet.

We also adjust the API description of this field to make it more
accurate.