Use `Tx.txOutMaxCoin` constant to replace `MinimumUTxO.maxLengthCoin`.
Although the `txOutMaxCoin` constant is lower than the `maxLengthCoin`
constant, it is still greater than or equal to any `Coin` value that can
appear within a transaction output.

As a result of this commit, the golden minimum UTxO values do not change
in any way.

If the `txOutMaxCoin` constant is ever changed, then this change will be
tested by `MinimumUTxOSpec`.