Add function `applyNM`.
This commit adds the `applyNM` function and property tests.

This function applies a monadic function @n@ times to an initial value,
returning the final result.

Its behaviour is similar to `applyN`, but applies a monadic function
instead of a pure function, hence the addition of the `M` suffix:

```
applyN  :: forall   a.            Int -> (a ->   a) -> a ->   a
applyNM :: forall m a. Monad m => Int -> (a -> m a) -> a -> m a
```