Adjust signature of `prop_computeMinimumCoinForUTxO_CardanoApi_CardanoLedger`.
It's simpler for this property to accept a `TokenBundle` instead of a
`TokenMap` and `Coin`.

The `Arbitrary` instance for `TokenBundle` uses the `genTxOutTokenBundle`
generator, which already has good coverage of minimum and maximum values
specifically in the context of a transaction output.

This allows us to get rid of the `Arbitrary` instance for `Coin` (which
in hindsight would have been better to implement with the `genTxOutCoin`
generator).

In response to review feedback:

https://github.com/input-output-hk/cardano-wallet/pull/3456#discussion_r953637703