cluster: double the epochLength
In #3461 we are seeing the cluster setup fail. It seems the chain gets
broken. A pool must produce a block within 3k/f = previously 75 slots
or the chain will be broken, with "NoLedgerView" being a symptom.
(https://input-output-rnd.slack.com/archives/CR599HMFX/p1649430846682959?thread_ts=1649430803.174879&cid=CR599HMFX)

We do see TraceNoLedgerView in the logs of the CI failures:

[pool-1:cardano.node.Forge:Error:34] [2022-08-25 09:29:47.00 UTC] fromList [("credentials",String "Cardano"),("val",Object (fromList [("kind",String "TraceNoLedgerView"),("slot",Number 30.0)]))]
[pool-1:cardano.node.LeadershipCheck:Info:34] [2022-08-25 09:29:47.20 UTC] {"chainDensity":0,"credentials":"Cardano","delegMapSize":4,"kind":"TraceStartLeadershipCheck","slot":31,"utxoSize":5263}
[pool-1:cardano.node.Forge:Error:34] [2022-08-25 09:29:47.20 UTC] fromList [("credentials",String "Cardano"),("val",Object (fromList [("kind",String "TraceNoLedgerView"),("slot",Number 31.0)]))]

/however/ the first TraceNoLedgerView we see is at slot 30, well before
slot 75.

Regardless, this commit increases the epoch length in hope that it might
alleviate the issue.

Changing the epochLength /could/ lead to other problems, but e.g.
`waitForNextEpoch` should still work correctly with a 90s timeout
waiting at most for 32s, so hopefully we're good.