Specialise all `Coin` <-> `Quantity` conversions to "lovelace".
In practice, when using the `Quantity` type to represent lovelace
quantities, we only ever instantiate the first type variable to
"lovelace", and never instantiate it to anything else.

Therefore, we can afford to be a bit more restrictive in the types of
our functions that convert between `Coin` and `Quantity`, limiting them
to conversions of "lovelace" (and not something else that we don't know
how to safely convert).