Remove the so-called "pointless" tests from `Api.TypesSpec`.
It's not clear what value these tests provide.

Of course, they do allow us to artificially increase our coverage from
the point of view of the code coverage checker. But it's not clear why
we should want to do this.

Moreover, artificially increasing the coverage score in this way may
hide real gaps in test coverage: because of these "pointless" tests, our
code coverage report currently can't tell us which types are really
exercised by our test suite, and which are not.

Additionally, the set of types tested here is incomplete: many newer
types were never added to the list.

If we really need to have this kind of test, then instead of maintaining
these test definitions manually (a process that is error prone, requires
a lot of boilerplate, and increases review overhead), it would be better
to generate them automatically: then we can automatically cover all
types used in the API, rather than just an subset.

However, if our goal is to reach 100% coverage of all API record field
accessors, perhaps it would be better to focus our efforts on writing
meaningful tests that actually stimulate the API.