Parameterise `UtxoTooSmall` by `ApiErrorTxOutputLovelaceInsufficient`.
WARNING: This commit changes the encoding of nullary `ApiError`
constructors from JSON strings to tagged JSON objects. This is
absolutely NOT what we want, as it will break compatibility with
existing clients that expect JSON strings for nullary constructors.

(See changes to `ApiErrorCode.json` for details.)

NOTE: Future commits will rectify this problem and restore compatibility
with existing clients.

Example:

Before this commit, the encoding for `WalletAlreadyExists` was:
```
"wallet_already_exists"
```

After this commit, the encoding for `WalletAlreadyExists` is:
```
{"tag": "wallet_already_exists"}
```