Add JSON encoding helper type `TaglessSum`.
To demonstrate the behaviour of this helper type, we add a JSON
roundtrip encoding test for `TaglessSum ApiErrorCode`.

This gives us the tagless encoding we want for nullary constructors of
the `ApiErrorCode` type.

However, in the case of non-nullary constructors of `ApiErrorCode`, this
encoding is NOT what we want: for those we want to keep the tag, so that
when decoding we can unambiguously discriminate between one error type
and another.

Therefore, for the final encoding of `ApiErrorCode`, we might need to
use some combination of `TaglessSum` and `DefaultSum`.