Tweak retrospective computation of change & outs
Without this change integration tests would fail e.g. where:

```pseudo-code
constructTx
  paymentTo addr1 (0 ada)

  |
  |
  V

Tx
  output: addr1 (1 ada) -- min ada, set automatically
  output: addr2 (9 ada) -- change
```

as we'd believe both outputs in the resuling tx were change -- neither
of them exists in the original output list.

If there are 'n' outputs in the original unbalanced tx, we will with
this commit instead judge that the first 'n' are "outputs" and that the
ones after that are all change. While this makes assumptions about how
balanceTx appends outputs, it is more importantly ressilient to outputs
being modified, like with the `increaseZeroAdaOutputs` feature.