Create module `Control.Monoid.Random.NonRandom`.
The `NonRandom` data type provides a stateless context for computations
that must be fully-deterministic, but where the function performing the
computation requires a `MonadRandom` instance.

Coin selection currently depends on the `NonRandom` data type as part of
its verification of post-conditions. We'd like to depend on this type
without depending on the `Control.Monad.Random.Extra` module.