Drop `cabal build` check in buildkite CI
This check is broken in the node bump and I don't know how to make it green.
Regardless, it seems fragile (often requiring caches to be cleared) and
will stop working in a few days when the packet.net agents become
unavailible.

The ability to build the project through haskell.nix is checked by hydra/cicero.

The `cabal build` checks were added during a time when `stack` was the
main tool for development in cardano-wallet. Now we only use `cabal`. If
the "cabal in nix-shell" build were to break separately from the nix
build, we'd notice fast enough.