Move macOS builds from bors to `master`
With this change, macOS unit tests and builds are no longer required to
merge a PR into master. However, they are now run on `master`, i.e.
/after/ a PR is merged.

To be sure a PR doesn't break mac builds ahead of merging, we can still
manually unblock the steps in buildkite. This could be useful for node
bumps etc.