Add function `shrinkBoundedEnum` to `Test.QuickCheck.Extra`.
Though the `QuickCheck` library provides a handy `arbitraryBoundedEnum`
function, it currently doesn't provide a `shrinkBoundedEnum`
counterpart. Users have asked for a function like this here:

https://github.com/nick8325/quickcheck/issues/343

This commit adds our own version, for now. Perhaps, later on, we can
submit a PR to the `QuickCheck` repository to add this function, and
delete our own version.