Pass `Maybe ScriptTemplate` to `balanceTransaction`
This is a quick and dirty solution to get balanceTx working with
constructSharedTransaction. Coin-selection needs to be aware how many
witnesses per input are expected.

It may be nicer to pass in some kind of `newtype CoinSelection = ...` to
`balanceTransaction` where it's up to the caller to have an appropriate
coin-selection given the the wallet type and `UTxO`.

It might also be that we could identify the upper bound of witnesses per
input purely based on the `UTxO`.