Add top-level `ErrBalanceTxBalanceInsufficient` error.
This record type is a copy of the coin selection `BalanceInsufficient`
error, but is specific to the `balanceTransaction` function.

Once all wallet endpoints have been rewritten to use
`balanceTransaction` as their main point of entry for balancing
transactions, we will be free to vary the internal API between
`balanceTransaction` and coin selection modules.