Provide HTTP server support for `ErrBalanceTxBalanceInsufficient`.
The server handler is an exact duplicate of the handler for the internal
coin selection error `BalanceInsufficient`, with exactly the same
message string.

Unfortunately, we're not yet in a position to remove the old error
handler, as even after `balanceTransaction` is converted to use this
error, there will still be some wallet endpoints that call into coin
selection modules directly, and thus the old error handler will still
be reachable.

But once all endpoints have been converted to use `balanceTransaction`,
we can remove this handler.