Make `-fno-warn-ambiguous-fields` conditional on the GHC version.
It's not possible to disable this type of warning in GHC 8.10.7, because
this type of warning was not yet invented at the time GHC 8.10.7 was
released.

However, we do need to disable these warnings in GHC 9.2.1
(temporarily), so we must make the "disabling of this warning type"
conditional on the compiler version.

We will eventually fix these warnings. See the following tracking
ticket:

https://input-output.atlassian.net/browse/ADP-2841