Restrict signing tests to recent eras
Quick workaround for failure in the Shelley era. The proper solution is
to also restrict the implementation to only work in recent eras, for
which a JIRA task was created.

Techincally, hiding this failure is a API breaking change, but it would
be /highly/, highly unlikely for any user to care about signing txs in
the Shelley era... with non-compatibile timelock scripts. Besides, the
preceding constructTx endpoint only works in recent eras.