Strengthen `prop_applyBlocks_filteredTxs_someOurs`.
The `applyBlocks` function was recently changed so that fees of incoming
transactions, if present (i.e., not `Nothing`), are no longer stripped
away (i.e., no longer replaced with `Nothing`).

The following commits are relevant:

120c7decdc662607758de60c7c8690845baeb22b
634265eb4bb13bc0c7782c889753affce62e1760

Consequently, it is now more straightforward to test our expectations
about the `fee` field within our property tests for `applyBlocks`. In
particular, we now test that `applyBlocks` does not modify the `fee`
field in any way (if it is not `Nothing`), whereas before, the `fee`
field was deliberately excluded from the test.

Note that this does rely on the fact that our transaction sequence
generator never generates transactions with fees that are `Nothing`:

https://github.com/input-output-hk/cardano-wallet/blob/4af4c184a2f986a51c44c0f93a5eef7e7fa31acf/lib/wallet/src/Cardano/Wallet/Primitive/Types/Tx/TxSeq/Gen.hs#L320

If in future we revise this generator to also cover transactions where
fees can be `Nothing`, we will have to also revise these property tests.

This commit strengthens the `prop_applyBlocks_filteredTxs_someOurs`
property, which tests the `applyBlocks` function in scenarios where
**_only some_** transactions in an arbitrary block sequence are
relevant to an arbitrary wallet.