Delete `TRANSMETA_ESTIMATE_01a` and `TRANSMETA_ESTIMATE_01b`.
The integration test suite is arguably the wrong place to test that fee
estimates should increase or decrease if the outputs of a transaction
are changed.

Moreover, depending on the initial starting conditions and the random
selection performed for each round of fee estimation, it's not
guaranteed that if the outputs for transaction `a` are smaller than the
outputs for transaction `b`, then the fee estimate for transaction `a`
will be lower than the fee estimate for transaction `b`.