Fix the restoration benchmark.
The `estimate tx fee` section of the restoration benchmark is currently
failing with errors similar to:

```hs
ExceptionSelectAssets
  (ErrSelectAssetsSelectionError
  (SelectionOutputErrorOf
  (SelectionOutputError
  { outputIndex = 0
  , outputErrorInfo =
    SelectionOutputCoinInsufficient
      (SelectionOutputCoinInsufficientError
        { minimumExpectedCoin = Coin 995610
        , output =
          ( Address ".."
          , TokenBundle {coin = Coin 1, tokens = TokenMap (fromList [])}
          )
        }
      )
    })))
```

This error occurs when a user-specified output has a lovelace quantity
that is greater than zero, but lower than the minimum value that would
be acceptable to the ledger. In this case, `balanceTx` will refuse to
balance the transaction, returning an error similar to the one above.
(Note that `1` lovelace is lower than `995610` lovelace, which is the
minimum accceptable amount of lovelace for this particular output.)

This commit changes the user-specified output to have a lovelace
quantity of zero. If an output has a lovelace quantity of zero, then
`balanceTx` (which is called by `transactionFee`) should automatically
assign a minimal quantity of lovelace to that output before balancing
the transaction and computing the fee.