Remove warning about exporting default constructor for `TokenMap`.
It's no longer necessary to worry about exporting the default
constructor for `TokenMap`.

The original concern was that exporting the default constructor would
allow callers to break the invariant that there are no `mempty` values
in the map.

However, this is no longer possible, as the `MonoidMap` type handles
this invariant automatically.

Therefore, it's no longer necessary to have a warning.

Nevertheless, this commit doesn't export the default constructor, as
there currently isn't any need to do so.