Drop unneeded `TxWitnessByronIcarusUTxO`
The differentiation between TxWitnessByronUTxO and TxWitnessByronIcarusUTxO
isn't needed. Where we case-switch the logic appears to be exactly the
same.

Context and motivation: At a high level we're moving from

```
Proxy k -> TxWitnessTag
```

(modulo constraints) to

```
Proxy k -> UTxOAssumptions -> TxWitnessTag
```

where `UTxOAssumptions` treats Byron and Icarus the same, and maps it to
the same `TxWitnessTag`.

By showing the byron-icarus differentiation in `TxWitnessTag` is unneeded
we can be assured that it is also unneeded in `UTxOAssumptions`, and we
can get by without adding a new Icarus-specific `UTxOAssumptions` constructor.