Drop unused generators and conversions in Write.Tx
The idea behind them was to make it easy for the wallet API to represent
tx outputs in their full form with inline scripts and datums included.

This is currently unused, and is starting to show signs of neglect (when
to use 'era' and when 'LatestLedgerEra'?). Probably it makes sense to
define binary/JSON conversions in a separate module anyway. Possibly it
also makes sense to base this upon a third module where ledger types are
redefined without the `era` type parameters.

Let's re-add this when needed.