Revise name of `increaseZeroAdaOutputs` function.
This function name seems a bit unclear and/or ambiguous.

It could be read as "increases the number of zero ada outputs", which is
obviously not what we intend.

This commit adjusts the function name to be explicit about the fact
that we're assigning minimal ada quantities to outputs without ada,
which is much less ambiguous.