Add `ErrBalanceTxAssetsInsufficientError`.
This is a clone of the `BalanceInsufficientError` type from
`coin-selection`.

Additional documentation has been added to explain this error in the
context of the `balanceTx` API.

In the near future, we can change the types of these fields to
ledger-compatible types (as used elsewhere in `balanceTx`), without
needing to change `cardano-coin-selection` (which may still rely on
wallet primitive types for the medium-term future).