Move `Compatibility.{Ledger,LedgerSpec}` to `conversions` package.
The `Compatibility.Ledger` module was located in `cardano-balance-tx`.
The `Compatibility.LedgerSpec` module was located in `cardano-wallet`.

This commit co-locates both modules within a new package called
`cardano-wallet-conversions`.

This makes it possible for both `cardano-wallet` and
`cardano-balance-tx` to depend upon these conversion functions without
having to export them from the public interface of `cardano-balance-tx`.

Note that in order to perform this relocation, this commit also clones
the definition of a single QuickCheck generator:
`Cardano.Wallet.Gen.genScript`.