Clone the test utility function `cardanoTx`.
This test utility function is simple enough that it doesn't matter too
much if we duplicate its definition across test suites.

In this particular case, the duplication allows us to reduce the number
of exports from `Write.Tx.BalanceSpec` to `TransactionSpec`, and brings
us one step closer to being able to move `BalanceSpec` to the
`cardano-balance-tx` library.