Clone `mockCardanoApiPParamsForBalancing` to `TransactionSpec`.
This will allow these mock parameters to be used by the tests for
`txConstraints` without also requiring `TransactionSpec` to depend on
`Write.Tx.BalanceSpec`.

Note that this clone is not 100% identical to the original: in
particular, we simplify the definition of `protocolParamCostModels`.
This is fine, because the `txConstraints` function does not use this
parameter.

We can therefore avoid cloning `costModelsForTesting`.