Remove `Write.withConstraints era` function
And favor the constraint `IsRecentEra era` over the
value `RecentEra era`.

Now that we use `PParams era` instead of `PParams (ShelleyLedgerEra era)`
we can often infer the era, just requiring the `IsRecentEra` constraint.

And now all the constraints we are interested in are superclass
constraints of `IsRecentEra`, so we don't need `withConstraints` to
bring them into scope.