Use qualified import style for functions relating to `Range` type.
Now that the `Range` type lives within its own module, there's no need
for every function in the `Range` module to include the name `Range`.

Instead, we can adopt a more minimal naming scheme similar to that used
by the `containers` library, and arrange that functions from the `Range`
module are imported into a `Range` namespace.